+++ internal/provider/model_sdwan_custom_control_topology_policy_definition.go
@@ -42,6 +42,7 @@ type CustomControlTopologyPolicyDefinitionSequences struct {
 	Name          types.String                                                  `tfsdk:"name"`
 	Type          types.String                                                  `tfsdk:"type"`
 	IpType        types.String                                                  `tfsdk:"ip_type"`
+	BaseAction    types.String                                                  `tfsdk:"base_action"`
 	MatchEntries  []CustomControlTopologyPolicyDefinitionSequencesMatchEntries  `tfsdk:"match_entries"`
 	ActionEntries []CustomControlTopologyPolicyDefinitionSequencesActionEntries `tfsdk:"action_entries"`
 }
@@ -131,6 +132,9 @@ func (data CustomControlTopologyPolicyDefinition) toBody(ctx context.Context) st
 			}
 			if !item.IpType.IsNull() {
 				itemBody, _ = sjson.Set(itemBody, "sequenceIpType", item.IpType.ValueString())
+			}
+			if !item.BaseAction.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "baseAction", item.BaseAction.ValueString())
 			}
 			if len(item.MatchEntries) > 0 {
 				itemBody, _ = sjson.Set(itemBody, "match.entries", []interface{}{})
@@ -320,6 +324,11 @@ func (data *CustomControlTopologyPolicyDefinition) fromBody(ctx context.Context,
 				item.IpType = types.StringValue(cValue.String())
 			} else {
 				item.IpType = types.StringNull()
+			}
+			if cValue := v.Get("baseAction"); cValue.Exists() {
+				item.BaseAction = types.StringValue(cValue.String())
+			} else {
+				item.BaseAction = types.StringNull()
 			}
 			if cValue := v.Get("match.entries"); cValue.Exists() {
 				item.MatchEntries = make([]CustomControlTopologyPolicyDefinitionSequencesMatchEntries, 0)
@@ -576,6 +585,9 @@ func (data *CustomControlTopologyPolicyDefinition) hasChanges(ctx context.Contex
 				hasChanges = true
 			}
 			if !data.Sequences[i].IpType.Equal(state.Sequences[i].IpType) {
+				hasChanges = true
+			}
+			if !data.Sequences[i].BaseAction.Equal(state.Sequences[i].BaseAction) {
 				hasChanges = true
 			}
 			if len(data.Sequences[i].MatchEntries) != len(state.Sequences[i].MatchEntries) {