+++ internal/provider/resource_sdwan_custom_control_topology_policy_definition_test.go
@@ -40,6 +40,7 @@ func TestAccSdwanCustomControlTopologyPolicyDefinition(t *testing.T) {
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.name", "Region1"),
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.type", "route"),
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.ip_type", "ipv4"),
+					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.base_action", "accept"),
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.match_entries.0.type", "ompTag"),
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.match_entries.0.omp_tag", "100"),
 					resource.TestCheckResourceAttr("sdwan_custom_control_topology_policy_definition.test", "sequences.0.action_entries.0.type", "set"),
@@ -63,6 +64,7 @@ resource "sdwan_custom_control_topology_policy_definition" "test" {
 		name = "Region1"
 		type = "route"
 		ip_type = "ipv4"
+		base_action = "accept"
 		match_entries = [{
 			type = "ompTag"
 			omp_tag = 100